src/app/root/root.component.spec.ts
import { TestBed, async } from '@angular/core/testing';
import { RootComponent } from './root.component';
import { SidebarComponent } from '../components/sidebar/sidebar.component';
import { MainContentComponent } from '../components/main-content/main-content.component';
import { OverlayModule} from '../components/overlay/overlay.module';
describe('AppComponent', () => {
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports:[OverlayModule],
      declarations: [
        RootComponent,
        SidebarComponent,
        MainContentComponent
      ],
    }).compileComponents();
  }));
  it('should create the app', async(() => {
    const fixture = TestBed.createComponent(RootComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  }));
  it(`should have as title 'app'`, async(() => {
    const fixture = TestBed.createComponent(RootComponent);
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('app');
  }));
  it('should render title in a h1 tag', async(() => {
    const fixture = TestBed.createComponent(RootComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('Welcome to app!');
  }));
});
